+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Tariff Zones added and mapped with stop places. These include the
+  defined geographic polygon as GML posList.
+- Topographic places benefited from the same import as above and now
+  have the GML polygon area available.
 
 ### Changed
 - Supports Laravel 9.x
+
+### Fixed
+- Stop place import now properly counts available elements before
+  processing, giving a correct progress bar.
 
 ## [0.1.2] – 2022-06-30